project_ABORA/BackEnd/requirements.txt
ag2==0.9.1.post0
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
autogen==0.9.1.post0
bcrypt==4.0.1
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
openai==1.77.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
regex==2024.11.6
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
termcolor==3.1.0
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.4.0
uvicorn==0.34.0
